feat: Implement dynamic event listener with threading
- Refactored the entire architecture into a modular, multi-file system (agent_core, agent_logic, listener, config, tools).
- Implemented a background listener using Python's `threading` to watch for external events without blocking the main chat.
- Created a file-based notification system (`/notifications`) to simulate asynchronous events.
- Differentiated agent workflows based on invocation source (user vs. autonomous listener) to manage human-in-the-loop approvals securely.
- Resolved race conditions for console I/O using `threading.Lock`.
- Implemented an archive system for agent replies (`/sent_replies`) to prevent feedback loops.
